# Generate RSA-SHA1 signature using openssl