--- a/aws/cloudfront/custom_policy.go
+++ b/aws/cloudfront/custom_policy.go
@@ -2,10 +2,7 @@ package cloudfront
 
 import (
 	"bytes"
-	"crypto"
-	"crypto/rsa"
 	"encoding/json"
-	"io"
 	"strings"
 	"time"
 )
@@ -87,18 +84,7 @@ func (p CustomPolicy) String() string {
 }
 
 // signWithPrivateKey returns a binary encoding of the Custom Policy signature
-func (p CustomPolicy) signWithPrivateKey(privateKey *rsa.PrivateKey) ([]byte, error) {
-	// create a sha1 digest of our policy
-	hashFunc := crypto.SHA1
-	h := hashFunc.New()
-	io.WriteString(h, p.String())
-	digest := h.Sum(nil)
-
+func (p CustomPolicy) signWithPrivateKey(privateKey PrivateKey) ([]byte, error) {
 	// calculates the signature of digest using RSASSA-PKCS1-V1_5-SIGN from RSA PKCS#1 v1.5.
-	// NOTE: By passing in nil instead of rand.Reader here, we disable RSA blinding.
-	if signature, err := rsa.SignPKCS1v15(nil, privateKey, hashFunc, digest); err != nil {
-		return []byte{}, err
-	} else {
-		return signature, nil
-	}
+	return privateKey.SignPKCS1v15([]byte(p.String()))
 }