Due to performance reason (see the benchmark output below)
we are switching from go crypto library to openssl.
Since we just have to deal with RSA-SHA1 signatures, we
implemented a minimum set of bindings which in theory make compilation
straightforward and should work with any recent version of openssl.

This is a breaking change.

Go 1.4.1 crypto library [old]
Go 1.4.1 openssl bindings [new]

    benchmark                     old ns/op     new ns/op     delta
    BenchmarkSignCustomPolicy     7956337       2742651       -65.53%

Is interesting to notice that go devel show a nice improvement
over go1.4.1 crypto. Probably due math/big optimizations.
Still not enough to match openssl but promising.

Go 1.4.1 crypto library [old]
Go devel +391805b Tue Feb 17 23:17:12 2015 +0000 crypto library [new]

    benchmark                     old ns/op     new ns/op     delta
    BenchmarkSignCustomPolicy     7956337       6291889       -20.92%